+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 新特性
 * 【poi】          增加ExcelUtil.getWriterWithSheet方法（感谢@【长沙】NULL）
 * 【core】        EnumUtil和ObjectUtil增加方法（pr#57@Gitee）
+* 【core】        EnumUtil增加fromString重载支持默认值（issue#IZFXJ@Gitee）
 
 ### Bug修复
 * 【core】         修复无效的日志打印（issue#IZFW9@Gitee）